Move hgCat, hgId, and hgStatus to hgCommandAsync.
This considerably speeds up atom while typing.
moving hgVersion and getHgPathStatus was also considered but it prooved to be quite hard.

With this change, the only synchronous commands happen when opening a project, and when saving a file.
I think this should be good enough given that saving a file usually calls a linter or formatter and it
takes some time anyways.

Related to: victor-torres/atom-hg#15